opengl: Fix loading unnamed GLES
Normally Android loads libGLES_${ro.hardware.egl}.so
But up to Android 14, it was allowed not to set ro.hardware.egl
In which case, Android would just load whatever libGLES_xxxx.so

Android 14.0 did it more finely, and would fallback only when
ro.hardware.egl is not set, which was still fine for older platforms.
Android 14.1 broke it by refusing the fallback when ro.board.platform is
set (which often is)

Simply allow the fallback in all cases.

Test:
- Blackview BV9800 Pro fails to boot without this patch
- It does boot with it